Initial commit
This small repo documents my process for building cross-platform ExpressJS Docker containers. I do this fairly often so it should be kept fairly up to date.

The code contained in this repo is a barebones-nodejs app with a dockerfile. I use this to ensure Buildx is working. The app has a single route that sends "Hello World" to localhost:8080.

I mainly use this process when developing express apps on amd64 and deploying them to arm64 and arm/v7 for use on Raspberry Pi.